Tweaks
Added more colors, cause colors are the shit.
Grouped the homebrew install in with the other apps. New if else statement inside the for loop to handle that.
Added a check to make sure script is ran as sudoer. With all output streams being redirected, non sudoer user would not suffice anymore. I think maybe cause it was prompting for input from the user.
Added another function to call the other functions. Makes it easier to modify the flow at the top of the script. Called at the bottom.
Alphabetized the apps to be installed so the app launcher would look cleaner.
Added Chrome and Atom so users would have more options.
Added terminal to the dock.