Moved to Python
Moving to python instead of bash. More options and more flexibility.

Bash script works for the most part except for the file encrypting and decrypting. I need to adjust which file gets created and which it saves too. It will also need to work on a loop system...

Create initial file
Append to it
Encrypt it
To view or add more...
Decrypt the file
Add
Encrypt again