Update
Broke the files up to have a new file generate the keys instead of one large script. Functions are imported.

Logic is broken on the main passGen.py file. Cannot view or gen the file twice without errors.

Current flow needs to have a text file prior to appending. May be able to fix this with a loop.